FieldCreation component
