+++ frontend/public/components/Dashboard.vue
@@ -7,64 +7,7 @@
             <TaskList></TaskList>
         </div>
         <div v-if="currTab == 'myForms' && !formPreview" id="form-creation">
-            <div id="field-input-container" v-if="!formSaved">
-                <p>Add a field:</p>
-                <p>
-                    Type:
-                    <select id="field-dropdown" v-model="currField">
-                        <option v-bind:value="field" v-for="field in fieldTypes" v-bind:key="field">{{caps(field)}}</option>
-                    </select>
-                </p>
-                <p>
-                    Label: 
-                    <input type="text" id="field-label">
-                </p>
-                <ul v-if="currField == 'dropdown'">
-                    <li>
-                        <button id="dropdown-add" @click="dropdownAdd">Add Dropdown Option</button>
-                    </li>
-                    <li v-for="option in [...Array(numDropdownFields).keys()]" v-bind:key="option">
-                        <input type="text" class="currDropdownOption" placeholder="New Dropdown Option">
-                    </li>
-                </ul>
-                <ul v-if="currField == 'multi-select'">
-                    <li>
-                        <button id="dropdown-add" @click="multiSelectAdd">Add Selection Option</button>
-                    </li>
-                    <li v-for="selection in [...Array(numMultiSelectFields).keys()]" 
-                    v-bind:key="selection"
-                    draggable 
-                    @dragstart='onDragMultiSelect(field)' 
-                    @dragenter='onDragEnterMultiSelect($event)' 
-                    @dragover.prevent 
-                    @dragenter.prevent
-                    >
-                        <input type="checkbox" class="multi-select-input">
-                        <input type="text" class="multi-select-label" placeholder="New Item Label">
-                        <i class="fa fa-bars clickable"></i>
-                    </li>
-                </ul>
-                <div v-if="currField == 'slider'">
-                    <p>
-                        Min: 
-                        <input type="number" id="slider-min">
-                    </p>
-                    <p>
-                        Max
-                        <input type="number" id="slider-max">
-                    </p>
-                    <p>
-                        Step
-                        <input type="number" id="slider-step">
-                    </p>
-                    <p>
-                        Initial
-                        <input type="number" id="slider-initial">
-                    </p>
-                </div>
-                <br>
-                <button id="field-add" @click="fieldAdd">Add</button>
-            </div>
+            <FieldCreation></FieldCreation>
             <div id="form-container">
                 <div id="form-content" v-if="!formSaved">
                     <input placeholder="Form Title" type="text" id="form-title" autocomplete="off" v-model="formTitle">
@@ -107,6 +50,7 @@
     import Calendar from '../components/Calendar.vue';
     import Inbox from '../components/Inbox.vue';
     import TaskList from '../components/TaskList.vue';
+    import FieldCreation from '../components/FieldCreation.vue';
     import FormPreview from '../components/FormPreview.vue';
 
     export default {
@@ -115,15 +59,12 @@
             Calendar,
             Inbox,
             TaskList,
+            FieldCreation
             FormPreview
         },
         data () {
             return {
                 currTab: "dashboard",
-                currField: '',
-                fieldTypes: ['text','number','text (long)','date','dropdown','multi-select','slider'],
-                numDropdownFields: 0,
-                numMultiSelectFields: 0,
                 fields: [],
                 inbox: [],
                 taskList: [],
@@ -255,12 +196,6 @@
                 this.draggingPos = targetPos;
             }
         },
-        onDragMultiSelect(item) {
-            return item
-        },
-        onDragEnterMultiSelect(event) {
-            return event
-        },
         updateInbox() {
             requests.getUserForms("ann_perkins").then(
                 response => {
@@ -281,44 +216,6 @@
                     }
                 }
             )
-        },
-        fieldAdd: function() {
-            let newField = {
-                label: document.getElementById("field-label").value,
-                type: document.getElementById("field-dropdown").value
-            };
-            if (this.currField == 'dropdown') {
-                newField.options = [];
-                for (let i of document.getElementsByClassName("currDropdownOption"))
-                    newField.options.push(i.value);
-                this.numDropdownFields = 0;
-                
-            }
-            else if (this.currField == "multi-select") {
-                newField.options = [];
-                for (let i of document.getElementsByClassName("multi-select-input"))
-                    newField.options.push(i.value);
-                this.numMultiSelectFields = 0;
-            }
-            else if (this.currField == 'slider') {
-                newField.slider = {};
-                newField.slider.min = document.getElementById("slider-min").value;
-                newField.slider.max = document.getElementById("slider-max").value;
-                newField.slider.step = document.getElementById("slider-step").value;
-                newField.slider.initial = document.getElementById("slider-initial").value;
-            }
-
-            this.fields.push(newField);
-
-            document.getElementById("field-dropdown").value = "";
-            this.currField = "";
-            document.getElementById("field-label").value = "";
-        },
-        dropdownAdd: function() {
-            this.numDropdownFields++;
-        },
-        multiSelectAdd: function() {
-            this.numMultiSelectFields++;
         },
         caps: function(text) {
             if (!text)
